Create render.yaml Blueprint
https://render.com/docs/infrastructure-as-code?_gl=1*1vbfcah*_ga*NDAwNDA5ODM0LjE3NDI1MDY5NjM.*_ga_QK9L9QJC5N*MTc0Mjc0NjA0OC43LjEuMTc0Mjc0NjA0OS4wLjAuMA..
A Blueprint acts as the single source of truth for configuring an interconnected set of services, databases, and environment groups. Whenever you update a Blueprint, Render automatically redeploys any affected services to apply the new configuration (you can disable this).

As your infrastructure grows over time, Blueprints become more and more helpful for managing changes and additions to it.